refactor(sandbox): i18n.ts を i18n/locale.ts へ移動し同名並存を解消
src/i18n.ts（ファイル）と src/i18n/（ディレクトリ）が同名で並存し、
import パス "../i18n" がどちらを指すのか読み手に曖昧だったため、
i18n ディレクトリへ統合する。

- 中身は locale 管理（locales 定義・resolveLocale・applyLocale・
  永続化・initializeI18n）なのでファイル名は locale.ts とする
- barrel file 禁止の運用のため index.ts は使わない
